+++ src/base/index.ts
@@ -53,7 +53,7 @@ export * from "./memo.js";
 export { Account } from "./account.js";
 export { MuxedAccount } from "./muxed_account.js";
 export type { TransactionSource } from "./transaction_source.js";
-export { Claimant } from "./claimant.js";
+export { Claimant, type HorizonPredicateJson } from "./claimant.js";
 export { Networks } from "./network.js";
 export { StrKey } from "./strkey.js";
 export { SignerKey } from "./signerkey.js";